requirements.txt
groq
python-telegram-bot
python-dotenv
streamlit